+++ src/nmath/gamma_cody.c
@@ -139,8 +139,8 @@ double attribute_hidden Rf_gamma_cody(double x)
 	.0057083835261 };
 
     /* Local variables */
-    long i, n;
-    long int parity;/*logical*/
+    int i, n;
+    int parity;/*logical*/
     double fact, xden, xnum, y, z, yi, res, sum, ysq;
 
     parity = (0);
@@ -187,7 +187,7 @@ double attribute_hidden Rf_gamma_cody(double x)
 	    /* -----------------------------------------------------------
 	       1 <= argument < 12, reduce argument if necessary
 	       ----------------------------------------------------------- */
-	    n = (long) y - 1;
+	    n = (int) y - 1;
 	    y -= (double) n;
 	    z = y - 1.;
 	}
